+++ src/Cell.js
@@ -1,4 +1,3 @@
-import {ENTITY_ID} from './const.js'
 import Fact from './Fact.js'
 
 /**
@@ -7,11 +6,5 @@ import Fact from './Fact.js'
  * The cells of the hypercube can be empty or full.
  * */
 export default class Cell extends Fact {
-	deleteProps(props) {
-		props.forEach(prop => {
-			if (prop !== ENTITY_ID) {
-				delete this[prop];
-			}
-		});
-	}
+
 }