+++ src/DimensionTable.spec.js
@@ -6,4 +6,7 @@ describe('class DimensionTable', function() {
 			new DimensionTable({});
 		}).toThrow();
 	});
+	it('generation unique entity ID from exist entities if they have one or more elements', () => {
+		expect(DimensionTable.reduceId([{id: 1}, {id: 3}], 'id')).toBe(4)
+	});
 });