+++ src/FactTable.primaryKey.spec.js
@@ -1,0 +1,33 @@
+import Cube from './Cube.js';
+import {isEqualObjects} from '../spec/helpers/helpers.js'
+
+describe('test primaryKey', () => {
+	it('fact table can have custom primaryKey', () => {
+		const factTable = {
+			facts: [
+				{ factId: 1, direction: 'left', date: '01' },
+				{ factId: 2, direction: 'right', date: '02' },
+			],
+			primaryKey: 'factId'
+		};
+		const dimensionHierarchies = [
+			{
+				dimensionTable: {
+					dimension: 'direction',
+					keyProps: ['direction']
+				},
+				dependency: [
+					{
+						dimensionTable: {
+							dimension: 'date',
+							keyProps: ['date']
+						},
+					}
+				]
+			}
+		];
+		const cube = Cube.create(factTable, dimensionHierarchies);
+		expect(cube.getCells()[0]["factId"]).toBeDefined();
+		isEqualObjects(cube.getFacts(), factTable.facts);
+	})
+});