+++ src/FactTable.spec.js
@@ -1,0 +1,7 @@
+import FactTable from './FactTable.js'
+
+describe('class FactTable', () => {
+	it('should define validateFactData', () => {
+		expect(FactTable.validateFactData).toBeDefined();
+	});
+});