+++ src/Member.js
@@ -1,5 +1,3 @@
-import {ENTITY_ID} from './const.js'
-
 /**
  * Element of dimension. Serving to determine the position and description of the data element
  * */
@@ -7,16 +5,16 @@ export default class Member {
 	constructor(data) {
 		Object.assign(this, data);
 	}
-	static create(id, props, data) {
+	static create(id, props, data, primaryKey) {
 		if (!(this === Member || Member.isPrototypeOf(this))) {
 			throw Error('this.constructor must be prototype of Member')
 		}
 		const memberData = {};
-		memberData[ENTITY_ID] = id;
+		memberData[primaryKey] = id;
 
 		props.forEach(prop => {
 			// исключить идентификатор самой сущности
-			if (prop !== ENTITY_ID) {
+			if (prop !== primaryKey) {
 				memberData[prop] = data[prop]
 			}
 		});