+++ src/const.js
@@ -1,2 +1,3 @@
-export const ENTITY_ID = 'id';
 export const DEFAULT_TEMPLATE_FOREIGN_KEY = '%s_id';
+export const DEFAULT_FACT_ID_PROP = 'id';
+export const DEFAULT_MEMBER_ID_PROP = 'id';